move IdGenerator behind lwfManager
